move all to outer folder, removing extra mars 511 folder
