+++ .gitignore
@@ -1,6 +1,20 @@
-.vscode/
-docs
-node_modules
-build
-npm-debug.log
-*.env
+# dependencies
+/node_modules
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
