--- mars511/.env.example
@@ -1,1 +1,0 @@
-NASA_API=""
