--- mars511/.vscode/tasks.json
@@ -1,15 +1,0 @@
-// Sass configuration
-{
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
-  "version": "2.0.0",
-  "tasks": [
-      {
-          "label": "Sass Compile",
-          "type": "shell",
-          "command": "node-sass src/css/main.scss src/css/main.css",
-          "group": "build",
-          "problemMatcher": []
-      }
-  ]
-}
