--- mars511/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "mars511",
-  "version": "0.1.0",
-  "description": "Creating a more pleasant commute on Mars",
-  "private": true,
-  "dependencies": {
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
-    "react-scripts": "1.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/digilou/mars511.git"
-  },
-  "keywords": [
-    "mars",
-    "space",
-    "NASA"
-  ],
-  "author": "Amy Carney",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/digilou/mars511/issues"
-  },
-  "homepage": "https://github.com/digilou/mars511#readme"
-}
