--- mars511/public/css/main.scss
@@ -1,132 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Open+Sans|Orbitron');
-
-/* Normalize */
-*{
-    margin:0;
-    padding:0;
-    box-sizing: border-box;
-}
-
-html {
-  height: 100%;
-}
-
-body{
-  background: rgb(255,0,0) url(/images/mars-bkgd.jpg) no-repeat center;
-  background-size: cover;
-  font-family: 'Open Sans', sans-serif;
-}
-
-/* Typography */
-
-h1{
-  color: rgb(254,188,72);
-  font: normal 3.5rem 'Orbitron', sans-serif;;
-  span {
-    display:block;
-    font-size:9rem;
-    font-style:italic;
-    font-weight:bold;
-    margin-top:-.35em;
-    margin-left:.12em;
-  }
-}
-
-label{
-  background-color:rgb(255,255,255);
-  margin:1rem -.3em;
-  padding:1rem;
-  font-size:2rem;
-  &:hover,
-  &:focus {
-    height:auto;
-    background: rgb(254,188,72);
-    font-weight:600;
-    color: rgb(0,0,0);
-  }
-}
-
-input[type="radio"]{
-  visibility:hidden;
-}
-
-/* Structure */
-footer {
-  display:block;
-  background-color: black;
-  color: lightgrey;
-  font-family: 'Orbitron', sans-serif;
-  margin-top:2.5rem;
-  line-height:1.5;
-  padding: 2rem 0 1rem 0;
-  position: relative;
-  bottom:0;
-  left:0;
-  width:100%;
-  bottom:0;
-  text-align: center;
-  border-top:.2em solid rgb(255,255,255);
-  li {
-    display:inline-block;
-    list-style-type:none;
-    padding:1rem .8rem;
-  }
-}
-
-footer a,
-footer li a{
-  color:rgb(255,0,0);
-}
-
-/* modules */
-#image-box{
-  background-color:rgba(255,255,255,.5);
-  border:1px solid #333;
-  box-shadow:5px 5px 5px rgb(0,0,0);
-  margin:0 auto;
-  width:20em;
-  height:15em;
-}
-#copyright{
-  font-size:small;
-}
-#forecast{
-  margin:0;
-  padding:.5em;
-  text-align:center;
-}
-#day{
-  display:block;
-}
-#day h2{
-  color:rgb(255,255,255);
-  font-size:4rem;
-  margin-bottom:0;
-}
-#day p{
-  color:rgb(255,255,255);
-  margin:0 auto;
-  font-size:2rem;
-}
-#day span{
-  font-weight:bold;
-}
-#temp{
-  display:block;
-  font-size:6rem;
-  font-weight:600;
-  color:rgb(255,0,0);
-  margin:1rem 0 2rem 0;
-}
-
-.active {
-  border: none;
-  background: rgb(254,188,72);
-  font-weight:600;
-  color: rgb(0,0,0);
-}
-
-#atmo{
-  margin:2rem 0;
-  font-size:3.5em;
-}
