--- mars511/public/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-us">
-  <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=Edge">
-    <meta name="author" content="Amy Carney">
-    <meta name="description" content="Take steps to make your commute on Mars more pleasant.">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" />
-    <link rel="stylesheet" href="%PUBLIC_URL%/css/main.min.css" type="text/css" />
-    <meta name="theme-color" content="#F00">
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-    
-    <title>Mars 511</title>
-  </head>
-  <body>
-    <noscript>
-      You need to enable JavaScript to run this app.
-    </noscript>
-    <div id="root"></div>
-  
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.0/jquery.min.js" crossOrigin="anonymous"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" crossOrigin="anonymous"></script>
-    <script src="scripts/api.js"></script>
-  </body>
-</html>
