--- mars511/public/offline.html
@@ -1,1 +1,0 @@
-<h1>Offline Skeleton</h1>
