+++ package.json
@@ -1,10 +1,18 @@
 {
-  "name": "mars-511",
-  "version": "1.0.0",
+  "name": "mars511",
+  "version": "0.1.0",
   "description": "Creating a more pleasant commute on Mars",
-  "main": "src/app.js",
+  "private": true,
+  "dependencies": {
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1",
+    "react-scripts": "1.1.4"
+  },
   "scripts": {
-    "start": "node src/server.js"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   },
   "repository": {
     "type": "git",
@@ -20,14 +28,5 @@
   "bugs": {
     "url": "https://github.com/digilou/mars511/issues"
   },
-  "homepage": "https://github.com/digilou/mars511#readme",
-  "dependencies": {
-    "express": "^4.16.3",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1"
-  },
-  "devDependencies": {
-    "gulp": "^3.9.1",
-    "gulp-sass": "^3.2.1"
-  }
+  "homepage": "https://github.com/digilou/mars511#readme"
 }
