+++ public/index.html
@@ -1,85 +1,27 @@
-<doctype html>
+<!DOCTYPE html>
 <html lang="en-us">
-
-<head>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=Edge">
-  <meta name="author" content="Amy Carney">
-  <meta name="description" content="Take steps to make your commute on Mars more pleasant.">
-  
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" />
-  <link rel="stylesheet" href="styles/main.css" />
-  
-  <title>Mars 511</title>
-</head>
-
-<body>
-  <noscript>
-    You need to enable JavaScript to run this app.
-  </noscript>
-  <div id="root"></div>
-
-  <!-- <header>
-    <h1>Mars<span>511</span></h1>
-  </header>
-
-  <main>
-   NASA image 
-    <section id="image-box">
-      NASA image of Mars 
-      <div id="image"></div>
-    </section>
-
-     
-        NASA image first;
-        Day:
-        Temp (toggle):
-        Atmospheric Conditions (icon?):
-        Humidity:
-        Wind Speed:
-        Wind Direction:
-        Pressure_string: (higher, lower)
-   
-
-
-    <section id="forecast">
-      <div id="day"></div>
-      <div id="temp"></div>
-        <label class="active" for="f">&deg;F</label>
-        <input type="radio" name="temp_select" id="f" checked>
-        <label for="c">&deg;C</label>
-        <input type="radio" name="temp_select" id="c">
-      <div id="atmo"></div>
-    </section>
-  </main>
+  <head>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=Edge">
+    <meta name="author" content="Amy Carney">
+    <meta name="description" content="Take steps to make your commute on Mars more pleasant.">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" />
+    <link rel="stylesheet" href="%PUBLIC_URL%/css/main.min.css" type="text/css" />
+    <meta name="theme-color" content="#F00">
+    <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
+    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
     
-  <footer>
-      <p>Weather API: <a href="http://marsweather.ingenology.com/">MAAS</a></p>
-      <p>Copyright &copy;2018 | Amy Carney</p>
-      <ul>
-      <li><a href="https://github.com/digilou">
-        <i class="fa fa-github fa-fw"></i>
-        </a>
-      </li>
-      <li><a href="https://www.linkedin.com/in/carneyamy">
-        <i class="fa fa-linkedin"></i>
-        </a>
-      </li>
-      <li><a href="https://carneydevelopoit.wordpress.com/">
-        <i class="fa fa-wordpress"></i>
-        </a>
-      </li>
-      <li><a href="https://www.freecodecamp.com/digilou">
-        (<i class="fa fa-fire fa-fw"></i>)
-        </a>
-      </li>
-      </ul>
-  </footer>
+    <title>Mars 511</title>
+  </head>
+  <body>
+    <noscript>
+      You need to enable JavaScript to run this app.
+    </noscript>
+    <div id="root"></div>
   
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.0/jquery.min.js" crossOrigin="anonymous"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" crossOrigin="anonymous"></script>
-  <script src="scripts/api.js"></script> -->
-
-</body>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.0/jquery.min.js" crossOrigin="anonymous"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" crossOrigin="anonymous"></script>
+    <script src="scripts/api.js"></script>
+  </body>
 </html>
