+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './public/css/index.min.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
